About text + fixups
Added the about text, and corrected the image coming down past it’s own
section when text is not long enough, with clearfix and clear: both.
Changed around the order of social media links and added Facebook link
in. Changed text on the submit button until it’s actually hooked up.